+++ app/(tabs)/plante/index.tsx
@@ -6,6 +6,7 @@ import { router } from 'expo-router';
 import { Ionicons } from '@expo/vector-icons';
 import { useEffect, useMemo, useState } from 'react';
 import { ScrollView, StyleSheet, Text, TextInput, View } from 'react-native';
+import { Header  } from "@/components/new/header/header";
 
 export default function Index() {
   const { vegetablesContext } = useVegetablesContext();
@@ -26,13 +27,7 @@ export default function Index() {
 
   return (
     <View style={styles.screen}>
-      <View style={styles.header}>
-        <View style={styles.headerLeft}>
-          <Ionicons name="leaf-outline" size={20} color="#FFFFFF" />
-          <Text style={styles.headerTitle}>Jardin</Text>
-        </View>
-        <Ionicons name="notifications-outline" size={20} color="#FFFFFF" />
-      </View>
+      <Header/>
 
       <ScrollView
         style={styles.container}
@@ -72,27 +67,6 @@ const styles = StyleSheet.create({
   screen: {
     flex: 1,
     backgroundColor: '#F3F4F6',
-  },
-  header: {
-    height: 92,
-    paddingTop: 44,
-    paddingHorizontal: 18,
-    backgroundColor: '#5A7F54',
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-  },
-  headerLeft: {
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 10,
-  },
-  headerTitle: {
-    fontSize: 18,
-    fontWeight: '700',
-    color: '#FFFFFF',
   },
 
   container: {
