+++ app/plan/index.tsx
@@ -2,16 +2,19 @@ import { BackButton } from "@/components/new/backButton/backButton";
 import { Plan } from "@/components/new/plan/plan";
 import { router } from "expo-router";
 import { StyleSheet, View } from "react-native";
-
+import { Header } from "@/components/new/header/header";
 
 export default function Index() {
 
 
     return (
         <View style={styles.container}>
+            <Header/>
+
             <View style={styles.backButton}>
                 <BackButton callback={() => router.replace('/home')} />
             </View>
+
             <Plan />
         </View>
     )
@@ -19,14 +22,13 @@ export default function Index() {
 
 const styles = StyleSheet.create({
     container: {
-        backgroundColor: '#F9FAFB',
-        paddingTop: 64,
-        paddingBottom: 50,
-        gap: 16,
-        height: '100%'
-    },
+    flex: 1,
+    backgroundColor: "#F3F4F6",
+  },
     backButton: {
-        zIndex: 3,
-        paddingLeft: 8,
+        paddingHorizontal: 16,
+        paddingTop: 12,
+        paddingBottom: 8,
+        zIndex: 1,
     }
 });
